Create Curl2_resonance.py
To solve for resonance solutions of the differential operator curl x curl using Python, we need to first define the eigenvalue equation associated with this operator. This can be done using the SciPy library, which provides a range of tools for solving eigenvalue problems. Here is some sample code that demonstrates how to solve for the resonances of curl x curl in a 2D domain: