Create Intergalactic_potential.ipynb
This code will generate a plot of the inter-galactic gravitational potential due to dark matter as a function of distance from the center of the dark matter halo. The x-axis represents the distance from the center of the dark matter halo in meters, while the y-axis represents the gravitational potential energy per unit mass in joules per kilogram. Note that the values of the parameters used in this example are for illustrative purposes only and may not accurately represent the actual values of the dark matter density and scale radius in the universe.