Create QC_F1D_int.ipynb
In this example, the function to be integrated is sin(x^2) * cos(y), which depends on two variables x and y. The Hamiltonian for two qubits is defined as a tensor product of two single-qubit Hamiltonians, where the frequency of the second qubit is a function of Omega and t, and the function to be integrated is passed as an argument. The initial state is defined as a tensor product of two qubits in the |00> state. The time and Omega arrays are defined, and the unitary matrix is calculated for each time step using the np.linalg.expm function. The final state is calculated by multiplying the final unitary matrix with the initial state. The integral is then calculated as the difference between the probability of measuring the |11> state in the final and initial states.

To perform numerical differentiation, we can use a similar approach, but instead of integrating the function over a range, we can use the gradient of the wave function to estimate the partial derivatives of the function with respect to each variable.